move fsck
